gui: change source-selector UI when tar file is selected
When tar file is selected, show file's path and "cancel" button. When
cancel button is clicked, clear the platformsList, the platform name etc
and disable buttons again.

TODO: I have avoided using redux' "Store" system since it seems (at
least at this stage) to be an overkill. Down the line, evaluate whether
this is the right way to go about it and fix (or don't) accordingly.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>